refactor: rename member and memberheader components
