+++ src/pages/manage/pages/members/sections/memberManagement/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
 import Button from "@/components/button/Button";
-import Member from "./Member";
-import MembersHeader from "./MembersHeader";
+import MemberTableRow from "./MemberTableRow";
+import MemberTableHead from "./MemberTableHead";
 import { useMemberManagement } from "./hooks/useMemberManagement";
 
 interface MemberManagementSectionProps {
@@ -29,35 +29,37 @@ const MemberManagementSection = ({
   }, [groupUuid]);
 
   return (
-    <div className="flex flex-col w-full justify-start items-start gap-[15px] overflow-x-hidden">
+    <div className="flex flex-col w-full justify-start items-start gap-[15px]">
       <div className="text-dark dark:text-grey text-[28px] font-bold">
         {t("manageGroup.members.list.title")}
       </div>
-      <div className="w-full overflow-x-scroll md:overflow-x-hidden">
-        <div className="min-w-full overflow-x-auto">
-          <table className="min-w-full table-fixed">
-            <MembersHeader />
-            <tbody>
-              {members.map((member) => (
-                <Member
-                  key={member.uuid}
-                  uuid={member.uuid}
-                  name={member.name}
-                  email={member.email}
-                  role={
-                    member.role === "admin"
-                      ? "관리자"
-                      : member.role === "manager"
-                        ? "매니저"
-                        : "일반"
-                  }
-                  onRoleChange={handleRoleChange}
-                />
-              ))}
-            </tbody>
-          </table>
-        </div>
+
+      <div className="w-full overflow-x-scroll overflow-y-visible">
+        <table className="min-w-[600px]">
+          <thead>
+            <MemberTableHead />
+          </thead>
+          <tbody>
+            {members.map((member) => (
+              <MemberTableRow
+                key={member.uuid}
+                uuid={member.uuid}
+                name={member.name}
+                email={member.email}
+                role={
+                  member.role === "admin"
+                    ? "관리자"
+                    : member.role === "manager"
+                      ? "매니저"
+                      : "일반"
+                }
+                onRoleChange={handleRoleChange}
+              />
+            ))}
+          </tbody>
+        </table>
       </div>
+
       <div className="flex justify-center self-stretch">
         <Button
           size="cta"
